+++ core/src/main/java/io/substrait/dsl/SubstraitBuilder.java
@@ -1507,7 +1507,24 @@ public class SubstraitBuilder {
    * @return a new {@link Type.UserDefined}
    */
   public Type.UserDefined userDefinedType(String urn, String typeName) {
-    return Type.UserDefined.builder().urn(urn).name(typeName).nullable(false).build();
+    return userDefinedType(urn, typeName, 0);
+  }
+
+  /**
+   * Creates a user-defined type with the specified URN, type name, and type variation.
+   *
+   * @param urn the URN of the extension containing the type
+   * @param typeName the name of the user-defined type
+   * @param typeVariationReference the type variation reference
+   * @return a new {@link Type.UserDefined}
+   */
+  public Type.UserDefined userDefinedType(String urn, String typeName, int typeVariationReference) {
+    return Type.UserDefined.builder()
+        .urn(urn)
+        .name(typeName)
+        .nullable(false)
+        .typeVariationReference(typeVariationReference)
+        .build();
   }
 
   // Misc