+++ core/src/main/java/io/substrait/expression/proto/ProtoExpressionConverter.java
@@ -289,6 +289,14 @@ public class ProtoExpressionConverter {
 
           return lambdaBuilder.lambdaFromStruct(parameters, () -> from(protoLambda.getBody()));
         }
+      case DYNAMIC_PARAMETER:
+        {
+          io.substrait.proto.DynamicParameter dp = expr.getDynamicParameter();
+          return Expression.DynamicParameter.builder()
+              .type(protoTypeConverter.from(dp.getType()))
+              .parameterReference(dp.getParameterReference())
+              .build();
+        }
       // TODO enum.
       case ENUM:
         throw new UnsupportedOperationException("Unsupported type: " + expr.getRexTypeCase());